+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ### Version 1.4.0:
 - Added new setting "openInFullScreenMode", which states if the calendar should be rendered in full screen mode.
+- The "Enable browser notifications" options in the "Configuration" dialog is now tabbed in slightly (shows which is belongs to).
 - 
 
 ### Version 1.3.3: